Modified variables to long where needed to accomodate Gb size genomes

+++ src/main/java/com/javafxapp/Call.java
@@ -10,13 +10,13 @@ public class Call {
     private int length;
     private String chromosome;
     private int start;
-    private int absoluteStart;
+    private long absoluteStart;
     private int end;
     private String alternate;
     private String id;
     private Double alleleFreq;
     HashMap<String,String> genotypes = new HashMap<>();
-    public Call(String type, int length, String chromosome, int start, int absoluteStart, String alternate, String id, HashMap<String,String> genotypes) {
+    public Call(String type, int length, String chromosome, int start, long absoluteStart, String alternate, String id, HashMap<String,String> genotypes) {
         this.type = type;
         if (Objects.equals(type, "DEL")) {
             this.length = Math.abs(length);
@@ -53,7 +53,7 @@ public class Call {
         return this.chromosome;
     }
 
-    public int getAbsoluteStart() {
+    public long getAbsoluteStart() {
         return this.absoluteStart;
     }
